+++ Classes/Nakama/Public/NakamaSDK/NTopicPresence.h
@@ -29,10 +29,10 @@ namespace Nakama {
 	public:
 		NTopicPresence() {}
 		NTopicPresence(TopicPresence message) : topic(NTopicId(message.topic())) {
-			for (size_t i = 0; i < message.joins_size(); i++) {
+			for (int i = 0; i < message.joins_size(); i++) {
 				join.push_back(NUserPresence(message.joins(i)));
 			}
-			for (size_t i = 0; i < message.leaves_size(); i++) {
+			for (int i = 0; i < message.leaves_size(); i++) {
 				leave.push_back(NUserPresence(message.leaves(i)));
 			}
 		}
